Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 
 int main(){
-    int a, b;
+    float a, b;
     char op;
-    scanf("%d %d %c", &a, &b, &op);
+    scanf("%f %f %c", &a, &b, &op);
     int result = a;
     if (op == '+'){
         printf("%d", a + b);
